refactor(core): use u16 for page_count
- using u16 for page count information is sufficient and removes
unnecessary casts on most places
- Pager uses u16

[no changelog]

+++ core/embed/rust/src/ui/component/base.rs
@@ -469,7 +469,7 @@ pub struct EventCtx {
     place_requested: bool,
     paint_requested: bool,
     anim_frame_scheduled: bool,
-    page_count: Option<usize>,
+    page_count: Option<u16>,
     button_request: Option<ButtonRequest>,
     root_repaint_requested: bool,
     swipe_disable_req: bool,
@@ -546,17 +546,17 @@ impl EventCtx {
         self.paint_requested
     }
 
-    pub fn set_page_count(&mut self, count: usize) {
+    pub fn set_page_count(&mut self, count: u16) {
         // #[cfg(feature = "ui_debug")]
         // assert!(self.page_count.unwrap_or(count) == count);
         self.page_count = Some(count);
     }
 
-    pub fn map_page_count(&mut self, func: impl Fn(usize) -> usize) {
+    pub fn map_page_count(&mut self, func: impl Fn(u16) -> u16) {
         self.page_count = Some(func(self.page_count.unwrap_or(1)));
     }
 
-    pub fn page_count(&self) -> Option<usize> {
+    pub fn page_count(&self) -> Option<u16> {
         self.page_count
     }
